Move line builder units in the unit directory
