+++ connector_ecommerce/unit/__init__.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
 from . import sale_order_onchange
+from . import line_builder
